+++ admin/view-posts.php
@@ -104,9 +104,9 @@
                                             <td><a href="view-comments.php"><?php echo $postCommentCount; ?></a></td>
                                             <td><?php echo $postDate; ?></td>
                                             <td class="text-center">
-                                                <a href="posts.php?source=view-post" class="btn btn-info btn-xs btn-flat"><i class="far fa-eye"></i></a>
-                                                <a href="posts.php?source=edit-post&postID=<?php echo $postID; ?>" class="btn btn-primary btn-xs btn-flat"><i class="far fa-edit"></i></a>
-                                                <a href="posts.php?delete=<?php echo $postID; ?>" class="btn btn-danger btn-xs btn-flat"><i class="far fa-trash-alt"></i></a>
+                                                <a href="posts.php?source=view-post" class="btn btn-info btn-xs btn-flat" title="View Post"><i class="far fa-eye"></i></a>
+                                                <a href="posts.php?source=edit-post&postID=<?php echo $postID; ?>" class="btn btn-primary btn-xs btn-flat" title="Edit Post"><i class="far fa-edit"></i></a>
+                                                <a href="posts.php?delete=<?php echo $postID; ?>" class="btn btn-danger btn-xs btn-flat" onclick="return confirm('Are you sure you want to delete this record?');" title="Remove Post"><i class="far fa-trash-alt"></i></a>
                                             </td>
                                         </tr>
                                         <?php
